+++ packages/runtime/src/index.ts
@@ -1,2 +1,3 @@
 export * from "./RuntimeEvent";
 export * from "./RuntimeHost";
+export * from "./RuntimeServiceKeys";